refactor: set layer as a normal package
