+++ geeservermap/__init__.py
@@ -2,7 +2,7 @@
 """TODO Missing docstring."""
 
 from . import helpers as helpers
-from .elements import layers as layers
+from . import layers as layers
 from .main import MESSAGES as MESSAGES
 from .map import Map as Map
